Delay load the new loader code / styling.
