+++ bundles.config.js
@@ -523,6 +523,13 @@ exports.extensionBundles = [
     type: TYPES.MISC,
   },
   {
+    name: 'amp-loader',
+    version: '0.1',
+    latestVersion: '0.1',
+    options: {hasCss: true},
+    type: TYPES.MISC,
+  },
+  {
     name: 'amp-mathml',
     version: '0.1',
     latestVersion: '0.1',
